refactor: 디렉토리 구조 재설계.
- 도메인 별 디렉토리 구조화 진행.
- dummy data를 폴더로 넣어 공통으로 쓰도록 함.
