+++ src/Article/css/ArticleHeader.css
@@ -1,0 +1,36 @@
+.article-header {
+    display: flex;
+    flex-direction: column;
+    align-items: baseline;
+    margin-top: 80px;
+    margin-left: 40px;
+    width: 80vw;
+    height: 100vh;
+  }
+  
+  .article-option-block {
+    display: flex;
+    flex-direction: row;
+  }
+  
+  .article-option {
+    margin: 0 3px;
+    padding: 3px;
+    width: 100px;
+    text-align: center;
+    border: 1px solid green;
+    border-width: 2px 4px;
+    border-radius: 40px;
+    background-color: #1bbc9b;
+    color: white
+  }
+  
+  .article-detail {
+    display: flex;
+    justify-content: space-between;
+    margin-left: 10px;
+  }
+  
+  .article-created-time{
+    margin-left: 30px;
+  }
