+++ src/data/articleInfo.json
@@ -1,0 +1,25 @@
+{
+    "options" : [
+        "a",
+        "b",
+        "c"
+    ],
+    "title" : "hello world",
+    "anonymous" : true,
+    "author" : {
+        "name" : "hello",
+        "image" : "https://hello.com",
+        "github" : "lxxjn0",
+        "blog" : "https://ell.com",
+        "description" : ""
+    },
+    "createdDate" : "20200411",
+    "content" : "",
+    "hashtags" : [
+        "abc",
+        "def"
+    ],
+    "favorites" : 1,
+    "wonders" : 1,
+    "clips" : 1
+}
